Ensure the ZipFile is closed in isZipFile() of ZipDexContainer
When calling isZipFile() in ZipDexContainer, the ZipFile would remain open if the file was in fact a zip file but the ZipFile object would then go out of scope thus creating a resource leak. This ensures that the ZipFile is
closed by adding a finally clause containing a close call at the end of the try catch block.