Some dex files are listed in the oat file, but aren't present in the oat/vdex
Oat version 138 introduced the concept of an "uncompressed" dex file, where
the dex file is listed in the oat file, but it's not actually present in
the oat or vdex files, instead, it's in the apk as per usual.

This change makes dexlib2 ignore these files, treating them as if they are
not a part of the oat file.